backend/training/requirements.txt
# Training dependencies for BridgeComm models

# MediaPipe Model Maker for gesture recognition training
mediapipe-model-maker>=0.2.1

# TensorFlow for emotion model training
tensorflow>=2.15.0

# Keras for model building
keras>=3.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Kaggle API for dataset download
kaggle>=1.5.16

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0
